refactor: ユーティリティ関数を別ファイルへ切り出し
